Move ImageView into package
This is in preparation for the GUI-centric workflow.
